requirements.txt
diff-match-patch==20121119
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.1
django-cors-headers==2.1.0
django-import-export==1.0.0
django-toolbelt==0.0.1
djangorestframework==3.7.7
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.1
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
PyYAML==3.12
static3==0.7.0
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
